Moved color logic to helper class
